x86_64: Add a new module and include a bitmap struct
Using a packed bitmap will be done routinely throughout the project. So
introduce a bitmap struct to allow clearing and setting a dynamically
allocated set of bits.

We still do not have a good abstraction to represent an accessible address
with a dynamic size. So for now this bitmap just wrap an address that is
provided as a "u64". Accessing it is unsafe because it is unknown whether
this address will cause a fault or not when accessed. All of this should be
cleaned up when we have a good abstraction for an address.