x86_64: vmx: Implement VMLAUNCH/VMRESUME
Introduce bare-minimum required functionality to execute VMLAUNCH/VMRESUME.
We introduce:

1- a guest registers structure that will be used to backup guest state.
2- (assembly) code to context switch between a host and a guest and
   save/restore state meanwhile.

The context switch code tries to execute code as if it is in a continuous
steam by ensuring that on a VMExit the host code executes from where it
left of. That generally makes VMLaunch look like a normal function call
from the hypervisor point of view. So code like this would work fine:

	/* do something */
	vmcs.run()
	/* do something else */