Add a test case for the instruction decoder
Include a test for the instruction decoder and add a new ap.s binary that
is used as part of the test.

'ap.s' is just some random assembly file that I use to bring up APs in my
rust kernel, so I just pulled it here to add some tests for the instruction
decoder.